+++ config/token.php
@@ -1,0 +1,11 @@
+<?php
+
+/*
+    token config 
+*/
+return [
+    /*
+            set token expiration time in days
+    */ 
+    'expires_at'=> env('TOKEN_EXPIRES_AT', 8),
+];